requirements.txt
Django==5.1.4
pillow==11.0.0
djangorestframework==3.15.2
requests==2.32.3